bp_bridge_hub_wococo
Trait BridgeHubWococoFinalityApi

pub trait BridgeHubWococoFinalityApi<Block: BlockT>: Core<Block> {
    // Provided method
    fn best_finalized(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Option<HeaderId<Hash, BlockNumber>>, ApiError> { ... }
}

API for querying information about the finalized chain headers.

This API is implemented by runtimes that are receiving messages from this chain, not by this chain’s runtime itself.

Provided Methods

fn best_finalized( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Option<HeaderId<Hash, BlockNumber>>, ApiError>

Returns number and hash of the best finalized header known to the bridge module.